Quick Changes and NOTES
Moves over to arrays.
QUICK WARNINGS:
1) The audio file is unable to open a file dialog without the button
being pressed. A workaround would be to force a dialog opening, but this
doesn't seem to be possible.
2) Songs will prematurely end about 5 seconds early due to the delay.
This is because the audio source buffer node automatically shuts off the
entire system once the data buffer for the song is exhausted. A
workaround is to start the song some time later using a different audio
source node, but I am not sure if this is OK with everyone.
3) To solve (1) without automatic pop-ups, it's necessary to delay the
game until the button is pressed and the file is given. Given the
current canvas we work with regarding the different tests, this is
highly difficult since the button appears somewhere else...